+++ src/core/config/loader.go
@@ -1,9 +1,29 @@
 package config
 
+// TODO: split this up... see the imports in the related files, need to put files into diff packages
+// TODO: not only this.. check recent tests and all other files... weird imports and missing structure
+
 import (
 	"fmt"
 	"os"
 )
+
+// Repository the repository for the git team config
+type Repository interface {
+	Query() (Config, error)
+}
+
+type StaticValueDataSource struct {
+}
+
+func NewStaticValueDataSource() StaticValueDataSource {
+	return StaticValueDataSource{}
+}
+
+func (ds StaticValueDataSource) Query() (Config, error) {
+	cfg := Load()
+	return cfg, nil
+}
 
 // ReadOnlyProperties read only properties of the config
 type ReadOnlyProperties struct {